Fix overflow check in add512()
The common use-case for add512 function is when a caller wants to store
a result in the same variable as a summand, say, add512(x,y,x).
In this case the overflow check is wrong.

        if ( (r->QWORD[i] < y->QWORD[i]) ||
             (r->QWORD[i] < x->QWORD[i]) )

If x is the same variable as r, the second comparision is
meaningless. This patch fixes it.

Signed-off-by: Kirill K. Smirnov <kirill.k.smirnov@gmail.com>